Sortable: Move sortable into widgets folder
